style(components/items): move cards to `/cards`
