+++ app/components/items/cards/item-top-card.skeleton.tsx
@@ -2,7 +2,7 @@
 
 import { FaStar } from 'react-icons/fa6'
 
-import { ItemImageSkeleton } from './item-image.skeleton'
+import { ItemImageSkeleton } from '../misc/item-image.skeleton'
 
 import { cn } from '@app/utils/cn'
 import { Skeleton } from '@app/components/ui/skeleton'
