+++ app/components/items/index.ts
@@ -1,10 +1,10 @@
-export * from './item-image'
-export * from './item-top-card'
+export * from './misc/item-image'
+export * from './cards/item-top-card'
 export * from './list/items-list'
 export * from './list/items-list-element'
 export * from './list/items-list-element.skeleton'
-export * from './item-image.skeleton'
+export * from './misc/item-image.skeleton'
 export * from './list/items-list.skeleton'
-export * from './item-top-card.skeleton'
-export * from './relative-time'
+export * from './cards/item-top-card.skeleton'
+export * from './misc/relative-time'
 export * from './genre/genre-chip'
